+++ src/base/lobby/gameFoundNotification.js
@@ -1,0 +1,9 @@
+wrap(() => {
+  onPage('Play', () => {
+    eventManager.on('getWaitingQueue', function gameFound() {
+      if (!fn.active()) {
+        fn.notify('Match found!');
+      }
+    });
+  });
+});